requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6
google-generativeai==0.3.2
python-dotenv==1.0.0
Pillow>=10.0.0
pydantic>=2.0.0
pydantic-settings>=2.5.0
httpx==0.25.2
python-json-logger==2.0.7
